requirements.txt
flask
flask-sqlalchemy
sqlalchemy
flask-login
pyserial
requests
pyjwt==1.7.1
gunicorn
flask-socketio
playsound
vext
vext.gi
pyglet
